+++ framework/servlet/src/main/java/is/codion/framework/servlet/EntityServiceFactory.java
@@ -29,7 +29,7 @@ import is.codion.framework.db.rmi.RemoteEntityConnection;
 public final class EntityServiceFactory implements AuxiliaryServerFactory<RemoteEntityConnection, ServerAdmin, EntityService> {
 
 	@Override
-	public EntityService createServer(Server<RemoteEntityConnection, ServerAdmin> server) {
+	public EntityService create(Server<RemoteEntityConnection, ServerAdmin> server) {
 		return new EntityService(server);
 	}
 }